Changed names to rust std convention
