+++ src/compute/aggregate/mod.rs
@@ -268,7 +268,7 @@ mod tests {
             .collect::<Primitive<i64>>()
             .to(DataType::Int64);
         // create an array that actually has non-zero values at the invalid indices
-        let c = arithmetics::add(&a, &b).unwrap();
+        let c = arithmetics::basic::add::add(&a, &b).unwrap();
         assert_eq!(Some((1..=100).filter(|i| i % 3 == 0).sum()), sum(&c));
     }
