+++ src/compute/arithmetics/basic/add.rs
@@ -1,0 +1,56 @@
+use std::ops::Add;
+
+use crate::{
+    array::{Array, PrimitiveArray},
+    compute::arity::{binary, unary},
+    error::{ArrowError, Result},
+    types::NativeType,
+};
+
+#[inline]
+pub fn add<T>(lhs: &PrimitiveArray<T>, rhs: &PrimitiveArray<T>) -> Result<PrimitiveArray<T>>
+where
+    T: NativeType + Add<Output = T>,
+{
+    if lhs.data_type() != rhs.data_type() {
+        return Err(ArrowError::InvalidArgumentError(
+            "Arrays must have the same logical type".to_string(),
+        ));
+    }
+
+    binary(lhs, rhs, lhs.data_type().clone(), |a, b| a + b)
+}
+
+#[inline]
+pub fn add_scalar<T>(lhs: &PrimitiveArray<T>, rhs: &T) -> PrimitiveArray<T>
+where
+    T: NativeType + Add<Output = T>,
+{
+    let rhs = *rhs;
+    unary(lhs, |a| a + rhs, lhs.data_type())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::array::Primitive;
+    use crate::datatypes::DataType;
+
+    #[test]
+    fn test_add() {
+        let a = Primitive::from(&vec![None, Some(6), None, Some(6)]).to(DataType::Int32);
+        let b = Primitive::from(&vec![Some(5), None, None, Some(6)]).to(DataType::Int32);
+        let result = add(&a, &b).unwrap();
+        let expected = Primitive::from(&vec![None, None, None, Some(12)]).to(DataType::Int32);
+        assert_eq!(result, expected)
+    }
+
+    #[test]
+    fn test_add_mismatched_length() {
+        let a = Primitive::from_slice(vec![5, 6]).to(DataType::Int32);
+        let b = Primitive::from_slice(vec![5]).to(DataType::Int32);
+        add(&a, &b)
+            .err()
+            .expect("should have failed due to different lengths");
+    }
+}
