+++ src/compute/arithmetics/basic/sub.rs
@@ -1,0 +1,55 @@
+use std::ops::{Neg, Sub};
+
+use crate::{
+    array::{Array, PrimitiveArray},
+    compute::arity::{binary, unary},
+    error::{ArrowError, Result},
+    types::NativeType,
+};
+
+#[inline]
+pub fn sub<T>(lhs: &PrimitiveArray<T>, rhs: &PrimitiveArray<T>) -> Result<PrimitiveArray<T>>
+where
+    T: NativeType + Sub<Output = T>,
+{
+    if lhs.data_type() != rhs.data_type() {
+        return Err(ArrowError::InvalidArgumentError(
+            "Arrays must have the same logical type".to_string(),
+        ));
+    }
+
+    binary(lhs, rhs, lhs.data_type().clone(), |a, b| a - b)
+}
+
+#[inline]
+pub fn sub_scalar<T>(lhs: &PrimitiveArray<T>, rhs: &T) -> PrimitiveArray<T>
+where
+    T: NativeType + Sub<Output = T>,
+{
+    let rhs = *rhs;
+    unary(lhs, |a| a - rhs, lhs.data_type())
+}
+
+#[inline]
+pub fn negate<T>(array: &PrimitiveArray<T>) -> PrimitiveArray<T>
+where
+    T: NativeType + Neg<Output = T>,
+{
+    unary(array, |a| -a, array.data_type())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::array::Primitive;
+    use crate::datatypes::DataType;
+
+    #[test]
+    fn test_subtract() {
+        let a = Primitive::from(&vec![None, Some(6), None, Some(6)]).to(DataType::Int32);
+        let b = Primitive::from(&vec![Some(5), None, None, Some(6)]).to(DataType::Int32);
+        let result = sub(&a, &b).unwrap();
+        let expected = Primitive::from(&vec![None, None, None, Some(0)]).to(DataType::Int32);
+        assert_eq!(result, expected)
+    }
+}
