+++ src/compute/arithmetics/decimal/sub.rs
@@ -34,10 +34,7 @@ use crate::compute::utils::combine_validities;
 /// the precision and scale is different, then an InvalidArgumentError is
 /// returned. This function panics if the subtracted numbers result in a number
 /// smaller than the possible number for the selected precision.
-pub fn subtract(
-    lhs: &PrimitiveArray<i128>,
-    rhs: &PrimitiveArray<i128>,
-) -> Result<PrimitiveArray<i128>> {
+pub fn sub(lhs: &PrimitiveArray<i128>, rhs: &PrimitiveArray<i128>) -> Result<PrimitiveArray<i128>> {
     // Matching on both data types from both arrays This match will be true
     // only when precision and scale from both arrays are the same, otherwise
     // it will return and ArrowError
@@ -77,7 +74,7 @@ pub fn subtract(
 /// InvalidArgumentError is returned. If the result from the sum is smaller
 /// than the possible number with the selected precision then the resulted
 /// number in the arrow array is the minimum number for the selected precision.
-pub fn saturating_subtract(
+pub fn saturating_sub(
     lhs: &PrimitiveArray<i128>,
     rhs: &PrimitiveArray<i128>,
 ) -> Result<PrimitiveArray<i128>> {
@@ -124,7 +121,7 @@ pub fn saturating_subtract(
 /// InvalidArgumentError is returned. If the result from the subtraction is
 /// smaller than the possible number with the selected precision then the
 /// function returns an ArrowError::ArithmeticError.
-pub fn checked_subtract(
+pub fn checked_sub(
     lhs: &PrimitiveArray<i128>,
     rhs: &PrimitiveArray<i128>,
 ) -> Result<PrimitiveArray<i128>> {
@@ -168,7 +165,7 @@ pub fn checked_subtract(
 /// and precision is adjusted to the largest precision and scale. If during the
 /// addition one of the results is smaller than the min possible value, the
 /// result precision is changed to the precision of the min value
-pub fn adaptive_subtract(
+pub fn adaptive_sub(
     lhs: &PrimitiveArray<i128>,
     rhs: &PrimitiveArray<i128>,
 ) -> Result<PrimitiveArray<i128>> {
@@ -250,7 +247,7 @@ mod tests {
         ])
         .to(DataType::Decimal(5, 2));
 
-        let result = subtract(&a, &b).unwrap();
+        let result = sub(&a, &b).unwrap();
         let expected = Primitive::from(&vec![
             Some(-11111i128),
             Some(11100i128),
@@ -266,7 +263,7 @@ mod tests {
     fn test_subtract_decimal_wrong_precision() {
         let a = Primitive::from(&vec![None]).to(DataType::Decimal(5, 2));
         let b = Primitive::from(&vec![None]).to(DataType::Decimal(6, 2));
-        let result = subtract(&a, &b);
+        let result = sub(&a, &b);
 
         if result.is_ok() {
             panic!("Should panic for different precision");
@@ -278,7 +275,7 @@ mod tests {
     fn test_subtract_panic() {
         let a = Primitive::from(&vec![Some(-99999i128)]).to(DataType::Decimal(5, 2));
         let b = Primitive::from(&vec![Some(1i128)]).to(DataType::Decimal(5, 2));
-        let _ = subtract(&a, &b);
+        let _ = sub(&a, &b);
     }
 
     #[test]
@@ -299,7 +296,7 @@ mod tests {
         ])
         .to(DataType::Decimal(5, 2));
 
-        let result = saturating_subtract(&a, &b).unwrap();
+        let result = saturating_sub(&a, &b).unwrap();
         let expected = Primitive::from(&vec![
             Some(-11111i128),
             Some(11100i128),
@@ -328,7 +325,7 @@ mod tests {
         ])
         .to(DataType::Decimal(5, 2));
 
-        let result = saturating_subtract(&a, &b).unwrap();
+        let result = saturating_sub(&a, &b).unwrap();
 
         let expected = Primitive::from(&vec![
             Some(-99999i128),
@@ -359,7 +356,7 @@ mod tests {
         ])
         .to(DataType::Decimal(5, 2));
 
-        let result = checked_subtract(&a, &b).unwrap();
+        let result = checked_sub(&a, &b).unwrap();
         let expected = Primitive::from(&vec![
             Some(-11111i128),
             Some(11100i128),
@@ -375,7 +372,7 @@ mod tests {
     fn test_subtract_checked_overflow() {
         let a = Primitive::from(&vec![Some(-99999i128)]).to(DataType::Decimal(5, 2));
         let b = Primitive::from(&vec![Some(1i128)]).to(DataType::Decimal(5, 2));
-        let result = checked_subtract(&a, &b);
+        let result = checked_sub(&a, &b);
 
         match result {
             Err(err) => match err {
@@ -394,7 +391,7 @@ mod tests {
         // -11099.9989 -> 9, 4
         let a = Primitive::from(&vec![Some(11_1111i128)]).to(DataType::Decimal(6, 4));
         let b = Primitive::from(&vec![Some(11111_11i128)]).to(DataType::Decimal(7, 2));
-        let result = adaptive_subtract(&a, &b).unwrap();
+        let result = adaptive_sub(&a, &b).unwrap();
 
         let expected = Primitive::from(&vec![Some(-11099_9989i128)]).to(DataType::Decimal(9, 4));
 
@@ -407,7 +404,7 @@ mod tests {
         // 11110.8889 -> 9, 4
         let a = Primitive::from(&vec![Some(11111_0i128)]).to(DataType::Decimal(6, 1));
         let b = Primitive::from(&vec![Some(1111i128)]).to(DataType::Decimal(5, 4));
-        let result = adaptive_subtract(&a, &b).unwrap();
+        let result = adaptive_sub(&a, &b).unwrap();
 
         let expected = Primitive::from(&vec![Some(11110_8889i128)]).to(DataType::Decimal(9, 4));
 
@@ -420,7 +417,7 @@ mod tests {
         // -00000.001  -> 8, 3
         let a = Primitive::from(&vec![Some(11111_11i128)]).to(DataType::Decimal(7, 2));
         let b = Primitive::from(&vec![Some(11111_111i128)]).to(DataType::Decimal(8, 3));
-        let result = adaptive_subtract(&a, &b).unwrap();
+        let result = adaptive_sub(&a, &b).unwrap();
 
         let expected = Primitive::from(&vec![Some(-00000_001i128)]).to(DataType::Decimal(8, 3));
 
@@ -433,7 +430,7 @@ mod tests {
         // 100.0000 -> 7, 4
         let a = Primitive::from(&vec![Some(99_9999i128)]).to(DataType::Decimal(6, 4));
         let b = Primitive::from(&vec![Some(-00_0001i128)]).to(DataType::Decimal(6, 4));
-        let result = adaptive_subtract(&a, &b).unwrap();
+        let result = adaptive_sub(&a, &b).unwrap();
 
         let expected = Primitive::from(&vec![Some(100_0000i128)]).to(DataType::Decimal(7, 4));
